Refactor types into a separate types.ts file.
