+++ tsconfig.json
@@ -18,7 +18,8 @@
       "@/components/*": ["components/*"],
       "@/configs/*": ["configs/*"],
       "@/constants/*": ["constants/*"],
-      "@/lib/*": ["lib/*"]
+      "@/lib/*": ["lib/*"],
+      "@/types/*": ["types/*"],
     }
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx"],
